+++ src/beast/math/statistic/DiscreteStatistics.java
@@ -53,7 +53,7 @@ public class DiscreteStatistics {
             }
         }
 
-        return m / (double) count;
+        return m / count;
     }
 
     /**
@@ -139,7 +139,7 @@ public class DiscreteStatistics {
             count = count - 1; // for ML estimate
         }
 
-        return var / (double) count;
+        return var / count;
     }
 
 
@@ -239,7 +239,7 @@ public class DiscreteStatistics {
      * @return variance of x (ML estimator)
      */
     public static double varianceSampleMean(double[] x, double mean) {
-        return variance(x, mean) / (double) x.length;
+        return variance(x, mean) / x.length;
     }
 
     /**
@@ -249,7 +249,7 @@ public class DiscreteStatistics {
      * @return variance of x (ML estimator)
      */
     public static double varianceSampleMean(double[] x) {
-        return variance(x) / (double) x.length;
+        return variance(x) / x.length;
     }
 
 
@@ -319,7 +319,7 @@ public class DiscreteStatistics {
         double minRange = Double.MAX_VALUE;
         int hpdIndex = 0;
 
-        final int diff = (int) Math.round(proportion * (double) x.length);
+        final int diff = (int) Math.round(proportion * x.length);
         for (int i = 0; i <= (x.length - diff); i++) {
             final double minValue = x[indices[i]];
             final double maxValue = x[indices[i + diff - 1]];
@@ -389,6 +389,6 @@ public class DiscreteStatistics {
             gm += Math.log(x[i]);
         }
 
-        return Math.exp(gm / (double) len);
+        return Math.exp(gm / len);
     }
 }
