+++ src/test/beast/beast2vs1/trace/NumberFormatter.java
@@ -64,7 +64,7 @@ public class NumberFormatter {
         cutoffTable = new double[sf];
         long num = 10;
         for (int i = 0; i < cutoffTable.length; i++) {
-            cutoffTable[i] = (double) num;
+            cutoffTable[i] = num;
             num *= 10;
         }
         decimalFormat.setMinimumIntegerDigits(1);
