+++ src/test/beast/evolution/operator/UniformIntegerOperatorTest.java
@@ -71,7 +71,7 @@ public class UniformIntegerOperatorTest extends TestCase {
 	            uniformOperator.proposal();
 	            Integer [] values = parameter.getValues();
 	            for (int k = 0; k < values.length; k++) {
-	                int j = (int)(double) values[k];
+	                int j = (int)values[k];
 	                count[k][j] += 1; 
 	            }
 	        }
